requirements.txt
# FastAPI and web
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0
scikit-image>=0.22.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# FAISS for similarity search
faiss-cpu>=1.7.4

# Dataset handling
pydicom>=2.4.0
h5py>=3.10.0

# Visualization and explainability
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
tqdm>=4.66.0
requests>=2.31.0
pyyaml>=6.0.1

# Logging and monitoring
tensorboard>=2.15.0
